requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
redis==5.2.1
celery==5.4.0
django_celery_beat==2.7.0
requests==2.32.3
django-cors-headers==4.6.0